requirements.txt
streamlit>=1.28.0
python-pptx>=0.6.21
PyYAML>=6.0
requests>=2.28.0
